+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased]
+### Added
+- Support for newer versions of Unity (#1191)
+
 ## [3.11.0]
 ### Added
 - Support for LZMA2 compression using XZ